Split FeverProducer into a couple of modules
