+++ app/models/qernel/fever_facade/producer_adapter.rb
@@ -5,6 +5,10 @@ module Qernel
     # Represents a Fever participant which will provide energy needed to meet
     # demand.
     class ProducerAdapter < Adapter
+      include Attributes
+      include CarrierHelpers
+      include Inject
+
       # Public: Returns an appropriate adapter class to represent the given
       # converter in Fever.
       #
@@ -31,26 +35,8 @@ module Qernel
       end
 
       def inject!
-        producer   = participant.producer
-        production = producer.output_curve.sum
-
-        @converter.demand = (production * 3600) / output_efficiency # MWh -> MJ
-
-        if production.positive?
-          full_load_hours = production / total_value(:heat_output_capacity)
-
-          @converter[:full_load_hours]   = full_load_hours
-          @converter[:full_load_seconds] = full_load_hours * 3600
-        end
-
-        if @converter.converter.groups.include?(:aggregator_producer)
-          demand = participant.demand
-          link   = @converter.converter.output(:useable_heat).links.first
-
-          link.share = demand.positive? ? production / demand : 1.0
-        end
-
-        inject_curve!(:output) { producer.output_curve.to_a }
+        inject_demand!
+        inject_curve!(:output) { participant.producer.output_curve.to_a }
         inject_input_curves!
       end
 
@@ -65,60 +51,12 @@ module Qernel
         end
       end
 
-      def producer_for_carrier(carrier)
-        participant.producer if input?(carrier)
-      end
-
-      # Public: Returns if the named carrier (a Symbol) is one of the inputs to
-      # the converter used by this adapter.
-      #
-      # Returns true or false.
-      def input?(carrier)
-        !@converter.converter.input(carrier).nil?
-      end
-
-      # Public: Creates a callable which takes a frame number and returns how
-      # much demand there is for a given carrier in that frame. Accounts for
-      # output losses.
-      #
-      # Returns a proc.
-      def demand_callable_for_carrier(carrier)
-        if (producer = producer_for_carrier(carrier))
-          efficiency = output_efficiency
-          ->(frame) { producer.source_at(frame) / efficiency }
-        else
-          ->(*) { 0.0 }
-        end
-      end
-
       # Public: Returns if this adapter has any units installed.
       def installed?
         @converter.number_of_units.positive?
       end
 
       private
-
-      def output_efficiency
-        slots = @converter.converter.outputs.reject(&:loss?)
-        slots.any? ? slots.sum(&:conversion) : 1.0
-      end
-
-      def input_efficiency
-        1.0 / @converter.converter.input(input_carrier).conversion
-      end
-
-      # Internal: The total capacity of the Fever participant in each frame.
-      #
-      # Returns an arrayish.
-      def capacity
-        return total_value(:heat_output_capacity) unless @config.alias_of
-
-        DelegatedCapacityCurve.new(
-          total_value(:heat_output_capacity),
-          aliased_adapter.participant.producer,
-          input_efficiency
-        )
-      end
 
       # Internal: The Fever participant is an alias of a producer in another
       # group; fetch it!
@@ -129,49 +67,6 @@ module Qernel
 
         alias_group.adapters.detect do |adapter|
           adapter.converter.key == @config.alias_of
-        end
-      end
-
-      def reserve
-        volume  = total_value { @converter.dataset_get(:storage).volume }
-        reserve = Merit::Flex::SimpleReserve.new(volume)
-
-        # Buffer starts full.
-        reserve.add(0, volume)
-
-        reserve
-      end
-
-      def share
-        link = @converter.converter.output(:useable_heat).links.first
-
-        if link.lft_converter.key.to_s.include?('aggregator')
-          link.lft_converter.output(:useable_heat).links.first.share
-        else
-          link.share
-        end
-      end
-
-      # Public: Determines the name of the main input carrier to the converter.
-      def input_carrier
-        return @input_carrier if @input_carrier
-
-        slots =
-          @converter.converter.inputs.reject do |slot|
-            slot.carrier.key == :ambient_heat
-          end
-
-        if slots.length != 1
-          raise 'Cannot determine input carrier for Fever producer: ' \
-                "#{@converter.key}"
-        end
-
-        @input_carrier = slots.first.carrier.key
-      end
-
-      def inject_input_curves!
-        inject_curve!(full_name: "#{input_carrier}_input_curve") do
-          Array.new(8760, &demand_callable_for_carrier(input_carrier))
         end
       end
     end
